refactor: move RadioValue type to Checkbox/types.ts
